docs(elements): centralize output renderer fixtures
